Add custom templates.
The standard slack template outputs the entire json string for the ecr
message which doesn't look great. It looks like this is because the
default template calls a function called format_resource which doesn't
exist in the ecr class in the c7n library. So I've added in a new one
which just displays the name.