Add missing SSL policy check.
I've added in the check for the missing SSL policy and I've added the
roles and policies to allow this. They're the same policies as the encrypted
bucket check but I've created new ones to avoid confusion later.

I've also had to put some fixes in because it was difficult to deploy to
other environments without changing the script file. Now the management account
number is passed in as a parameter to the script and this is used to
construct the SQS queue URL in the management account. Otherwise,
everything is deployed depending on which account credentials you use.

I've updated the README and the GitHub deploy workflow to show this.

The other problem was that the slack url was wrong so we weren't getting
any messages through. This looks like it was because we were originally
only storing the last part of the slack webhook, but now we're storing
the whole url so we don't need to reconstruct it in here.