DR2-1651 Add inspector notifications.
I've had to add in a custom template because there's no default pretty
print for inspector like there is for other services.

I've also updated the python script which builds the custodian yaml
because it was assuming there was always an email and a slack notify
action which is annoying. We don't want to send an email.

Ran terraform fmt